refactor : InventoryCommonBottom 컴포넌트 작성
