+++ src/app.features/inventory/components/CigaretteList.tsx
