+++ src/app.features/inventory/types/ciga.ts
@@ -1,0 +1,16 @@
+export type ChoType =
+	| '전체'
+	| 'ㄱ'
+	| 'ㄴ'
+	| 'ㄷ'
+	| 'ㄹ'
+	| 'ㅁ'
+	| 'ㅂ'
+	| 'ㅅ'
+	| 'ㅇ'
+	| 'ㅈ'
+	| 'ㅊ'
+	| 'ㅋ'
+	| 'ㅌ'
+	| 'ㅍ'
+	| 'ㅎ';
